Correct the error message and keep the command line information.
The command line was originally reported as user information on a PEAR Error object. During the
import to git the PEAR Error was exchanged with an exception and the command line provided as
the second argument to the exception. As strings in this position get reset the information will
be lost. I assume that concatenating the command line to the error might actually be what was
desired.